Refactor existing / new implementations, bump year
Bump the year numbers 2024->2025, 2025->2026, to reflect that the
existing implementation changed in 2025 and the new implementation
probably won't be tested until 2026.

Add the existing implementations as new files, matching the test
filenames:

* api/authentication_2025.py
* api/views/terms_accepted_user_2025.py (extracted from ./privaterelay.py)
